+++ sys-devel/gettext/gettext-0.14.1.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2004 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-devel/gettext/gettext-0.14.1.ebuild,v 1.19 2004/12/17 06:15:48 vapier Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-devel/gettext/gettext-0.14.1.ebuild,v 1.20 2004/12/30 03:56:20 j4rg0n Exp $
 
 inherit eutils gnuconfig toolchain-funcs mono libtool
 
@@ -10,7 +10,7 @@ SRC_URI="mirror://gnu/${PN}/${P}.tar.gz"
 
 LICENSE="GPL-2"
 SLOT="0"
-KEYWORDS="~alpha amd64 ~arm ~hppa ~ia64 ~mips ~ppc ~ppc64 ppc-macos ~s390 ~sh ~sparc ~x86"
+KEYWORDS="~alpha amd64 ~arm ~hppa ~ia64 ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~x86"
 IUSE="emacs nls"
 
 DEPEND="virtual/libc"
@@ -59,8 +59,11 @@ src_install() {
 	doexe gettext-tools/misc/gettextize || die "doexe"
 
 	# remove stuff that glibc handles
-	rm -f ${D}/usr/include/libintl.h
-	rm -f ${D}/usr/$(get_libdir)/libintl.*
+	if ! use ppc-macos; then
+		# Mac OS X does not provide these files.
+		rm -f ${D}/usr/include/libintl.h
+		rm -f ${D}/usr/$(get_libdir)/libintl.*
+	fi
 	rm -rf ${D}/usr/share/locale/locale.alias
 	# /usr/lib/charset.alias is provided by Mac OS X
 	use ppc-macos && rm -f ${D}/usr/lib/charset.alias
